Idea: Use Quarkus to manage shared objects
This change just adds the plumbing, it doesn't actually start to take advantage of dependency injection
There are a few changes that need to happen in other projects to take full advantage of this.  Some of the class init order changes
and tests were failing because of how the tests deal with `user.home`.  Usages of `user.home` must be lazy to support the tests.

Depends on: okta/okta-sdk-java#464