Detect the correct set of properties to use for JHipster applications
Now that the CLI supports Quarkus, JHipster + Quarkus apps need to use the Quarkus property set.

Detection of quarkus is based on finding `generator-jhipster-quarkus` in the package.json in the current directory.

Fallback is Spring based properties

In the future, adding other property types (i.e. Micronaut can be done the same way)